+++ src/CoreEx.Validation/Rules/CommonRule.cs
@@ -1,6 +1,5 @@
 ﻿// Copyright (c) Avanade. Licensed under the MIT License. See https://github.com/Avanade/CoreEx
 
-using System;
 using System.Threading;
 using System.Threading.Tasks;