+++ src/CoreEx.Validation/Rules/EntityRule.cs
@@ -34,6 +34,12 @@ namespace CoreEx.Validation.Rules
             if (context.Value == null)
                 return;
 
+            if (Validator is CommonValidator<TProperty> vp) // Common validators need the originating context for best results.
+            {
+                await vp.ValidateAsync(context, cancellationToken).ConfigureAwait(false);
+                return;
+            }
+
             // Create the context args.
             var args = context.CreateValidationArgs();